requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2018.10.15
chardet==3.0.4
configparser==4.0.2
discord.py==1.3.3
emoji==0.5.2
gspread==3.0.1
httplib2==0.11.3
idna==2.9
idna-ssl==1.1.0
multidict==4.7.5
mysqlclient==1.4.4
numpy==1.12.0
oauth2client==4.1.3
pandas==0.24.1
Pillow==5.2.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
python-dateutil==2.8.0
pytz==2019.2
requests==2.20.0
rsa==4.0
setproctitle==1.1.10
six==1.11.0
typing-extensions==3.7.4.2
urllib3==1.24.1
websockets==8.1
yarl==1.4.2
